# api/api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgDx API",
    "description": "AgDx API facilitates sharing of crop pest and disease data between platforms and/or by providing links between complementary ICT tools where appropriate.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0.0-beta1"
  },
...
  "servers": [
    {
      "url": "http://localhost:9100",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/pest/update/{id}": {
      "put": {
        "tags": [
          "Pest data"
        ],
        "summary": "Update pest record",
        "operationId": "updatePestData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
...
    "/api/v1/crop/pest/update/{id}": {
      "put": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Update crop pest record",
        "operationId": "updatePestData_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropPestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/disease/update/{id}": {
      "put": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Update crop disease record",
        "operationId": "updatePestData_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropDiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropDiseaseEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/pest/add": {
      "post": {
        "tags": [
          "Pest data"
        ],
        "summary": "Add new pest",
        "operationId": "addPestData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/pest/add": {
      "post": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Add new pest",
        "operationId": "addPestData_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropPestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/disease/add": {
      "post": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Add new crop pest record",
        "operationId": "addPestData_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropDiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropDiseaseEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
...
    "/api/v1/pest/{pestName}/scientific": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using scientific name",
        "operationId": "getPestByScientificName",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pest/{pestName}/common": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using common name",
        "operationId": "getPestByCommonName",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pest/order/{order}": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using order",
        "operationId": "getPestByOrder",
        "parameters": [
          {
            "name": "order",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pest/genus/{genus}": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using genus",
        "operationId": "getPestByGenus",
        "parameters": [
          {
            "name": "genus",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pest/family/{family}": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using family",
        "operationId": "getPestByFamily",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePestEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/pest/{pestName}/scientific": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get details of specific pest using its scientific name",
        "operationId": "getPestByScientificName_1",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/{pestName}/common": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get details of specific pest using its common name",
        "operationId": "getPestByCommonName_1",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PestEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "commonName": {
            "type": "string",
            "description": "Pest common name",
            "example": "Wooly apple aphid"
          },
          "scientificName": {
            "type": "string",
            "description": "Pest scientific name",
            "example": "Eriosoma lanigerum"
          },
          "order": {
            "type": "string",
            "description": "Pest order",
            "example": "Hemiptera"
          },
          "family": {
            "type": "string",
            "description": "Pest family",
            "example": "Aphididae"
          },
          "genus": {
            "type": "string",
            "description": "Pest genus",
            "example": "Eriosoma"
          },
          "category": {
            "type": "string",
            "description": "Pest family",
            "example": "Insect",
            "enum": [
              "Virus",
              "Bacteria",
              "Nematode",
              "Plant",
              "Insect"
            ]
          }
        }
      },
...
      "CropPestEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "countryCode": {
            "type": "string",
            "description": "Can be ISO2(KE) or ISO3(KEN)",
            "example": "KE"
          },
          "lat": {
            "type": "number",
            "description": "GPS lat",
            "example": 36.821945
          },
          "lon": {
            "type": "number",
            "description": "GPS lon",
            "example": -1.292066
          },
          "alt": {
            "type": "number",
            "description": "Identification Altitude above sea level",
            "example": 1200
          },
          "accuracy": {
            "type": "number",
            "description": "Accuracy of GPS location in meters",
            "example": 1.0
          },
          "geodeticDatum": {
            "type": "string",
            "description": "North America Datum or World geodetic coordinate system",
            "example": "NAD83"
          },
          "locationLevel1": {
            "type": "string",
            "description": "Province, district or region",
            "example": "Carroll, Ohio, United States"
          },
          "locationLevel2": {
            "type": "string",
            "description": "Municipality,district",
            "example": "Victoria"
          },
          "locationLevel3": {
            "type": "string",
            "description": "Administrative level i.e city,down",
            "example": "CITY"
          },
          "dateObserved": {
            "type": "string",
            "description": "Date when it was detected",
            "format": "date",
            "example": "2021-01-13"
          },
          "dateIdentified": {
            "type": "string",
            "description": "Date when it was identified",
            "format": "date",
            "example": "2021-01-13"
          },
          "dateRecorded": {
            "type": "string",
            "description": "Date when it was recorded into the system",
            "format": "date",
            "example": "2021-01-13"
          },
          "dateUpdated": {
            "type": "string",
            "description": "Date the record was updated",
            "format": "date",
            "example": "2021-01-13"
          },
          "dateUncertaintyInDays": {
            "type": "integer",
            "description": "Date Accuracy in Days",
            "format": "int32",
            "example": 1
          },
          "cropName": {
            "type": "string",
            "description": "Crop name",
            "example": "Cassava"
          },
          "cultivarName": {
            "type": "string",
            "description": "Crop cultivar",
            "example": "TMS 90257"
          },
          "pestCommonName": {
            "type": "string",
            "description": "Common pest name",
            "example": "Wooly apple aphid"
          },
          "pestScientificName": {
            "type": "string",
            "description": "Scientific pest name",
            "example": "Eriosoma lanigerum"
          },
          "pestNativity": {
            "type": "string",
            "description": "Pest nativity status",
            "example": "Native",
            "enum": [
              "Introduced",
              "Native"
            ]
          },
          "pestDetectionStatus": {
            "type": "string",
            "description": "Pest detection status",
            "example": "Detected",
            "enum": [
              "Undetected",
              "Detected"
            ]
          },
          "pestCategory": {
            "type": "string",
            "description": "Pest category",
            "example": "Insect",
            "enum": [
              "Virus",
              "Bacteria",
              "Nematode",
              "Plant",
              "Insect"
            ]
          },
          "pestHabitat": {
            "type": "string",
            "description": "Pest habitat",
            "example": "Forest, Soil"
          },
          "pestAbundance": {
            "type": "string",
            "description": "Pest abundance",
            "example": "Single plant"
          },
          "pestDensity": {
            "type": "string",
            "description": "Range of percent cover or description (low, medium, common, rare, etc.)",
            "example": "Common",
            "enum": [
              "Low",
              "Medium",
              "High",
              "Common",
              "Rare"
            ]
          },
          "pestTreatmentStatus": {
            "type": "string",
            "description": "management status of the pest at time of recording (Treated, Untreated)",
            "example": "Treated",
            "enum": [
              "Treated",
              "Untreated"
            ]
          },
          "pestVector": {
            "type": "string",
            "description": "Vector of the pest",
            "example": "Pied Crow"
          },
          "pestIdentificationMethod": {
            "type": "string",
            "description": "This can contain multiple identification methods (PCR,ELISA)",
            "example": "PCR"
          },
          "pestVerificationMethod": {
            "type": "string",
            "description": "This can contain multiple verification methods",
            "example": "Visual"
          },
          "pestRecordBasis": {
            "type": "string",
            "description": "The basis of the occurrence (Human observation)",
            "example": "Human observation"
          },
          "pestVerified": {
            "type": "boolean",
            "description": "Observed pest has been verified"
          },
          "severityNumberScale": {
            "type": "integer",
            "description": "Severity scale on an number level (5)",
            "format": "int32",
            "example": 5
          },
          "severityPercentageScale": {
            "type": "number",
            "description": "Severity scale on an percentage level",
            "example": 0.5
          },
          "severityRangeScale": {
            "type": "string",
            "description": "Severity scale on an range level",
            "example": "1-5"
          },
          "pestIncidence": {
            "type": "string",
            "description": "Number of measure for disease occurrence on subject population",
            "example": "50 Percent"
          },
          "visitType": {
            "type": "string",
            "description": "Purpose of visit to area for collected data",
            "example": "Invasive Species Survey"
          },
          "areaUnit": {
            "type": "string",
            "description": "Area measurement unit ha,acre,sqm",
            "example": "ha"
          },
          "sampledAreaSize": {
            "type": "number",
            "description": "Size of area sampled",
            "example": 20.5
          },
          "affectedAreaSize": {
            "type": "number",
            "description": "Size of area affected",
            "example": 15
          },
          "sampleSource": {
            "type": "string",
            "description": "Source of the sample i.e Field, greenhouse, lab",
            "example": "Field"
          },
          "sampleDesc": {
            "type": "string",
            "description": "Description of the sample i.e roots, leaves",
            "example": "Roots"
          },
          "numOfSamples": {
            "type": "integer",
            "description": "Number of samples collected",
            "format": "int32",
            "example": 500
          },
          "numOfAffectedSamples": {
            "type": "integer",
            "description": "Number of affected samples collected",
            "format": "int32",
            "example": 350
          },
          "approximateQuantity": {
            "type": "string",
            "description": "Approximate or range of subjects observed or biocontrol agents released (15, 15-20)",
            "example": "15"
          },
          "sampleUnits": {
            "type": "string",
            "description": "Unit of measure (individuals, adults, grams, etc.)",
            "example": "adults"
          },
          "labSampleId": {
            "type": "string",
            "description": "Lab sample identifier",
            "example": "SMPL-567"
          },
          "hasVisualSymptoms": {
            "type": "boolean",
            "description": "Specimens portrayed visual symptoms related to the pest"
          },
          "dataCollectionMethod": {
            "type": "string",
            "description": "Methodology/protocol used to take the data (Aerial survey)",
            "example": "Aerial survey"
          },
          "genbankAccessionNumber": {
            "type": "string",
            "description": "GeneBank number for identification and confirmation",
            "example": "LC.5656951"
          },
          "projectName": {
            "type": "string",
            "description": "Name of the project as the source of data",
            "example": "EDDMAPS"
          },
          "identifiedBy": {
            "type": "string"
          },
          "reportedBy": {
            "type": "string"
          },
          "recordedBy": {
            "type": "string"
          },
          "reviewedBy": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          },
          "identificationMethod": {
            "type": "string"
          },
          "visuallyIdentified": {
            "type": "boolean"
          },
          "originalSystemId": {
            "type": "string",
            "description": "In the case of data from an aggregator,the unique identifier from the original data set.Used in cases where two aggregators may encounter the same data."
          },
          "contributingSystemId": {
            "type": "string",
            "description": "Unique identifier from system contributing data for followup and updating"
          }
        }
      },
      "CropDiseaseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "countryCode": {
            "type": "string",
            "description": "Can be ISO2(KE) or ISO3(KEN)",
            "example": "KE"
          },
          "lat": {
            "type": "number",
            "description": "GPS lat",
            "example": 36.821945
          },
          "lon": {
            "type": "number",
            "description": "GPS lon",
            "example": -1.292066
          },
          "alt": {
            "type": "number",
            "description": "Identification Altitude above sea level",
            "example": 1200
          },
          "accuracy": {
            "type": "number",
            "description": "Accuracy of GPS location in meters",
            "example": 1.0
          },
          "geodeticDatum": {
            "type": "string",
            "description": "North America Datum or World geodetic coordinate system",
            "example": "NAD83"
          },
          "locationLevel1": {
            "type": "string",
            "description": "Province, district or region",
            "example": "Carroll, Ohio, United States"
          },
          "locationLevel2": {
            "type": "string",
            "description": "Municipality,district",
            "example": "Victoria"
          },
          "locationLevel3": {
            "type": "string",
            "description": "Administrative level i.e city,down",
            "example": "CITY"
          },
          "dateObserved": {
            "type": "string",
            "description": "Date when it was detected",
            "format": "date",
            "example": "2021-01-01"
          },
          "dateIdentified": {
            "type": "string",
            "description": "Date when it was identified",
            "format": "date",
            "example": "2021-01-13"
          },
          "dateRecorded": {
            "type": "string",
            "description": "Date when it was recorded into the system",
            "format": "date",
            "example": "2021-01-13"
          },
          "dateUpdated": {
            "type": "string",
            "description": "Date the record was updated",
            "format": "date",
            "example": "2021-01-13"
          },
          "cropName": {
            "type": "string",
            "description": "Crop name",
            "example": "Banana"
          },
          "cultivarName": {
            "type": "string",
            "description": "Crop cultivar",
            "example": "TMS 90257"
          },
          "pathogenCommonName": {
            "type": "string",
            "description": "Common pest name",
            "example": "BXW"
          },
          "pathogenScientificName": {
            "type": "string",
            "description": "Scientific pest name",
            "example": "Xanthomonas wilt of banana"
          },
          "pathogenNativity": {
            "type": "string",
            "description": "Pest nativity status",
            "example": "Introduced",
            "enum": [
              "Introduced",
              "Native"
            ]
          },
          "pathogenDetectionStatus": {
            "type": "string",
            "description": "Pathogen detection status",
            "example": "Detected",
            "enum": [
              "Undetected",
              "Detected"
            ]
          },
          "pathogenCategory": {
            "type": "string",
            "description": "Pathogen category",
            "example": "Virus",
            "enum": [
              "Virus",
              "Bacteria",
              "Fungi",
              "Protozoa",
              "Plants"
            ]
          },
          "pathogenVector": {
            "type": "string",
            "description": "Pathogen vector",
            "example": "Aphids, Equipment"
          },
          "pathogenIdentificationMethod": {
            "type": "string",
            "description": "This can contain multiple identification methods (PCR,ELISA)",
            "example": "PCR"
          },
          "pathogenTreatmentStatus": {
            "type": "string",
            "description": "management status of the pathogen at time of recording (Treated, Untreated)",
            "example": "Treated",
            "enum": [
              "Treated",
              "Untreated"
            ]
          },
          "pathogenVerificationMethod": {
            "type": "string",
            "description": "This can contain multiple verification methods",
            "example": "Visual"
          },
          "pathogenRecordBasis": {
            "type": "string",
            "description": "The basis of the occurrence (Human observation)",
            "example": "Human observation"
          },
          "pathogenVerified": {
            "type": "boolean",
            "description": "Observed pathogen has been verified"
          },
          "severityNumberScale": {
            "type": "integer",
            "description": "Severity scale on an number level (5)",
            "format": "int32",
            "example": 5
          },
          "severityPercentageScale": {
            "type": "number",
            "description": "Severity scale on an percentage level",
            "example": 0.5
          },
          "severityRangeScale": {
            "type": "string",
            "description": "Severity scale on an range level",
            "example": "1-5"
          },
          "pathogenIncidence": {
            "type": "string",
            "description": "Number of measure for disease occurrence on subject population",
            "example": "50 Percent"
          },
          "areaUnit": {
            "type": "string",
            "description": "Area measurement unit ha,acre,sqm",
            "example": "ha"
          },
          "sampledAreaSize": {
            "type": "number",
            "description": "Size of area sampled",
            "example": 20.5
          },
          "affectedAreaSize": {
            "type": "number",
            "description": "Size of area affected",
            "example": 15
          },
          "sampleSource": {
            "type": "string",
            "description": "Source of the sample i.e Field, greenhouse, lab",
            "example": "Field"
          },
          "sampleDesc": {
            "type": "string",
            "description": "Description of the sample i.e roots, leaves",
            "example": "Roots"
          },
          "numOfSamples": {
            "type": "integer",
            "description": "Number of samples collected",
            "format": "int32",
            "example": 500
          },
          "numOfAffectedSamples": {
            "type": "integer",
            "description": "Number of affected samples collected",
            "format": "int32",
            "example": 350
          },
          "sampleUnits": {
            "type": "string",
            "description": "Unit of measure (individuals, pieces, grams, etc.)",
            "example": "grams"
          },
          "labSampleId": {
            "type": "string",
            "description": "Lab sample identifier",
            "example": "SMPL-100"
          },
          "hasVisualSymptoms": {
            "type": "boolean",
            "description": "Specimens portrayed visual symptoms related to the disease"
          },
          "dataCollectionMethod": {
            "type": "string",
            "description": "Methodology/protocol used to take the data (Field visit)",
            "example": "Field visit"
          },
          "genbankAccessionNumber": {
            "type": "string",
            "description": "GeneBank number for identification and confirmation ofr the disease",
            "example": "LC.789546"
          },
          "projectName": {
            "type": "string",
            "description": "Name of the project as the source of data",
            "example": "CIAT"
          },
          "identifiedBy": {
            "type": "string"
          },
          "reportedBy": {
            "type": "string"
          },
          "recordedBy": {
            "type": "string"
          },
          "reviewedBy": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          },
          "identificationMethod": {
            "type": "string"
          },
          "visuallyIdentified": {
            "type": "boolean"
          },
          "originalSystemId": {
            "type": "string",
            "description": "In the case of data from an aggregator,the unique identifier from the original data set.Used in cases where two aggregators may encounter the same data."
          },
          "contributingSystemId": {
            "type": "string",
            "description": "Unique identifier from system contributing data for followup and updating"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api": {
        "type": "http",
        "in": "header",
        "scheme": "basic"
      }
    }
  }
}